ivshmem: reserve space for inter-VM synchronization data
First few KB should be used for lock and for segments, as all VMs should
have same view of which pieces of memory are already used.
For lock, std::atomic is used. Hopefuly it will work.
Guard pages are also added.

One-time initialization is needed. But only the very first VM is allowed
to do initial setup; later VM should not destroy existing data. Currently
setup involves only write guard pages. Lock doesn't need setup, it is just
a number. But segments data are written later, and that memory should not
be erased.

Signed-off-by: Justin Cinkelj <justin.cinkelj@xlab.si>